Move windrose into station dashboard and station dashboard folder
