+++ react/components/station_dashboard/station_dashboard.js
@@ -1,8 +1,15 @@
 import { FaWindowClose } from "react-icons/fa";
 import { empty_station_obj } from "../layout"
+import { getDisplayName } from "../utils/station_data_format_util";
+import { RenderWindRose } from "./wind_rose";
 
-export default function StationDashboard({children, selected_station, setSelectedStation}) {
-    console.log(children)
+export default function StationDashboard({children, selected_station, setSelectedStation, stationsDescriptions}) {
+    console.log(selected_station);
+    console.log(stationsDescriptions);
+    const station_name = selected_station[0];
+    const display_name = getDisplayName(stationsDescriptions, station_name);
+    const stationData = selected_station[1]["properties"]["station_data"]
+    console.log(stationData)
     return (
         <div key="01-station-dashboard" class="station_dashboard">
             <div className="dash-header">
@@ -12,11 +19,15 @@ export default function StationDashboard({children, selected_station, setSelecte
                         setSelectedStation(empty_station_obj)
                     }}
                 ><FaWindowClose/></button>
-                <h3>{selected_station[0]}</h3>
+                <h3>{display_name}</h3>
             </div>
             <div class="dash-body">
                 <p>BEHOLD! A Station dashboard goes here!</p>
-                <p>Here are the children in this tag!  {children}</p>
+                <p>
+                <RenderWindRose  
+                sourceData={stationData}
+                /> 
+                </p>
             </div>
             <div class="dash-footer">
                 <p>This the footer block</p>
